requirements.txt
blinker==1.9.0
certifi==2025.4.26
chardet==3.0.4
click==8.1.8
diskcache==5.6.3
Flask==3.1.0
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
itsdangerous==2.2.0
Jinja2==3.1.6
llama_cpp_python==0.3.8
MarkupSafe==3.0.2
numpy==2.2.5
PyAudio==0.2.14
pyttsx3==2.98
rfc3986==1.5.0
sniffio==1.3.1
SpeechRecognition==3.14.2
typing_extensions==4.13.2
Werkzeug==3.1.3
